+++ aasrp/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Initializing our tests
+"""